Updating for 1.22.x.  Harmony patch is correctly aimed at the new BlockEntitySoilNutrition.beginIntervalledUpdate method, but fails to actually patch.
